+++ src/onegov.directory/onegov/directory/archive.py
@@ -1,5 +1,4 @@
 import mimetypes
-import json
 import shutil
 import os
 
@@ -8,6 +7,7 @@ from onegov.core.csv import convert_list_of_dicts_to_csv
 from onegov.core.csv import convert_list_of_dicts_to_xlsx
 from onegov.core.csv import convert_xls_to_csv
 from onegov.core.csv import CSVFile
+from onegov.core.custom import json
 from onegov.core.utils import Bunch, rchop, is_subpath
 from onegov.directory.errors import MissingColumnError
 from onegov.directory.models import Directory, DirectoryEntry